requirements.txt
# --- API framework ---
fastapi==0.115.0
starlette==0.38.6
uvicorn[standard]==0.30.6
pydantic==2.8.2
httpx==0.27.2
openai==1.40.6

# --- Vector + math (pin numpy <2.0 for faiss) ---
numpy==1.26.4
faiss-cpu==1.8.0.post1

# --- Only needed if you still unpickle LangChain Document objects ---
# (safe to keep; helps when loading your .pkl bundles)
langchain-core==0.3.72
python-dotenv==1.0.1

